Refactor admission internals to prepare for new eviction policies
No functional change:

- Move `EntrySizeAndFrequency` from `sync::base_cache` and
  `future::base_cache` (duplicated) to `common::concurrent`
- Add `EvictionPolicyConfig::uses_frequency_sketch` instead of comparing
  against `EvictionPolicyConfig::TinyLfu` directly
- Fix doc wording in `sync::builder` ("to the cache" -> "of the cache")
