+++ src/sync/base_cache.rs
@@ -15,8 +15,8 @@ use crate::{
             deques::Deques,
             entry_info::EntryInfo,
             housekeeper::{Housekeeper, InnerSync},
-            AccessTime, KeyHash, KeyHashDate, KvEntry, OldEntryInfo, ReadOp, ValueEntry, Weigher,
-            WriteOp,
+            AccessTime, EntrySizeAndFrequency, KeyHash, KeyHashDate, KvEntry, OldEntryInfo, ReadOp,
+            ValueEntry, Weigher, WriteOp,
         },
         deque::{DeqNode, Deque},
         frequency_sketch::FrequencySketch,
@@ -828,29 +828,6 @@ impl EvictionCounters {
     }
 }
 
-#[derive(Default)]
-struct EntrySizeAndFrequency {
-    policy_weight: u64,
-    freq: u32,
-}
-
-impl EntrySizeAndFrequency {
-    fn new(policy_weight: u32) -> Self {
-        Self {
-            policy_weight: policy_weight as u64,
-            ..Default::default()
-        }
-    }
-
-    fn add_policy_weight(&mut self, weight: u32) {
-        self.policy_weight += weight as u64;
-    }
-
-    fn add_frequency(&mut self, freq: &FrequencySketch, hash: u64) {
-        self.freq += freq.frequency(hash) as u32;
-    }
-}
-
 // NOTE: Clippy found that the `Admitted` variant contains at least a few hundred
 // bytes of data and the `Rejected` variant contains no data at all. It suggested to
 // box the `SmallVec`.
@@ -1216,7 +1193,7 @@ where
                     self.apply_writes(&mut deqs, &mut timer_wheel, w_len, &mut eviction_state);
                 }
 
-                if self.eviction_policy == EvictionPolicyConfig::TinyLfu
+                if self.eviction_policy.uses_frequency_sketch()
                     && self.should_enable_frequency_sketch(&eviction_state.counters)
                 {
                     self.enable_frequency_sketch(&eviction_state.counters);
